Replace macloader with a shell script
This shell script will copy the MAC address found in EFS to the kernel
driver's sysfs. This allows to use the correct MAC address without
macloader.

The actual MAC address appears to exist in sec_efs; regular efs
partition will be used as a fallback if it does not exist in sec_efs.

As a side note, macloader itself never actually worked on gtowifi, as
Samsung never released the source code for the WLAN driver to actually
make use of it.

Also, move WiFi init scripts to device tree

Change-Id: I51d9b0bfab5032c9a6c321b3e86c48541551af1d